[refac] 리소스 조회 메서드를 util 패키지로 분리
