+++ cmd/get.go
@@ -1,13 +1,6 @@
 package cmd
 
 import (
-	"fmt"
-	"os"
-	"strconv"
-	"strings"
-
-	"github.com/dolong2/dcd-cli/api/exec"
-	"github.com/dolong2/dcd-cli/api/exec/response"
 	cmdError "github.com/dolong2/dcd-cli/cmd/err"
 	"github.com/dolong2/dcd-cli/cmd/resource"
 	"github.com/dolong2/dcd-cli/cmd/util"
@@ -36,186 +29,13 @@ var getCmd = &cobra.Command{
 			return cmdError.NewCmdError(1, "올바르지 않은 리소스 타입입니다.")
 		}
 
-		switch {
-		case resourceType.IsEqual(resource.Application):
-			err := getApplication(cmd)
-			if err != nil {
-				return err
-			}
-		case resourceType.IsEqual(resource.Workspace):
-			err := getWorkspace(cmd)
-			if err != nil {
-				return err
-			}
-		case resourceType.IsEqual(resource.ApplicationType):
-			err := printApplicationTypes()
-			if err != nil {
-				return cmdError.NewCmdError(1, err.Error())
-			}
-		case resourceType.IsEqual(resource.Domain):
-			err := getDomain(cmd)
-			if err != nil {
-				return cmdError.NewCmdError(1, err.Error())
-			}
-		case resourceType.IsEqual(resource.Env):
-			err := getEnv(cmd)
-			if err != nil {
-				return cmdError.NewCmdError(1, err.Error())
-			}
-		case resourceType.IsEqual(resource.VOLUME):
-			err := getVolume(cmd)
-			if err != nil {
-				return cmdError.NewCmdError(1, err.Error())
-			}
-		default:
-			return cmdError.NewCmdError(1, "조회할 수 없는 리소스 타입입니다.")
+		err := util.GetResource(cmd, resourceType)
+		if err != nil {
+			return cmdError.NewCmdError(1, err.Error())
 		}
 
 		return nil
 	},
-}
-
-func getWorkspace(cmd *cobra.Command) error {
-	id, existsFlagErr := cmd.Flags().GetString("id")
-	if id == "" || existsFlagErr != nil {
-		workspaceList, err := exec.GetWorkspaces()
-		if err != nil {
-			return cmdError.NewCmdError(1, err.Error())
-		}
-
-		usedWorkspaceId, usedWorkspace := util.GetWorkspaceId(cmd)
-		if usedWorkspace != nil {
-			usedWorkspaceId = ""
-		}
-
-		printWorkspaceList(workspaceList.List, usedWorkspaceId)
-
-		return nil
-	}
-
-	workspace, err := exec.GetWorkspace(id)
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	}
-
-	printWorkspace(*workspace)
-
-	return nil
-}
-
-func getApplication(cmd *cobra.Command) error {
-	workspaceId, err := util.GetWorkspaceId(cmd)
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	}
-
-	applicationId, err := cmd.Flags().GetString("id")
-
-	if applicationId == "" && err == nil {
-		labels, err := cmd.Flags().GetStringArray("label")
-
-		var applications *response.ApplicationListResponse
-
-		// id, labels 플래그 둘다 없을때, 조건 없이 애플리케이션 조회
-		if len(labels) == 0 || err != nil {
-			applications, err = exec.GetApplications(workspaceId)
-			if err != nil {
-				return cmdError.NewCmdError(1, err.Error())
-			}
-
-			printApplicationList(applications.Applications)
-
-			return nil
-		} else {
-			applications, err = exec.GetApplicationsByLabels(workspaceId, labels)
-			if err != nil {
-				return cmdError.NewCmdError(1, err.Error())
-			}
-		}
-
-		printApplicationList(applications.Applications)
-
-		return nil
-	}
-
-	application, err := exec.GetApplication(workspaceId, applicationId)
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	}
-
-	printApplication(*application)
-
-	return nil
-}
-
-func getDomain(cmd *cobra.Command) error {
-	workspaceId, err := util.GetWorkspaceId(cmd)
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	}
-
-	domainListResponse, err := exec.GetDomains(workspaceId)
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	}
-
-	printDomainList(domainListResponse.Domains)
-
-	return nil
-}
-
-func getEnv(cmd *cobra.Command) error {
-	workspaceId, err := util.GetWorkspaceId(cmd)
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	}
-
-	envId, err := cmd.Flags().GetString("id")
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	} else if envId == "" {
-		envListResponse, err := exec.GetEnvList(workspaceId)
-		if err != nil {
-			return cmdError.NewCmdError(1, err.Error())
-		}
-
-		printEnvList(*envListResponse)
-	} else {
-		envResponse, err := exec.GetEnv(workspaceId, envId)
-		if err != nil {
-			return cmdError.NewCmdError(1, err.Error())
-		}
-
-		printEnv(*envResponse)
-	}
-
-	return nil
-}
-
-func getVolume(cmd *cobra.Command) error {
-	workspaceId, err := util.GetWorkspaceId(cmd)
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	}
-
-	volumeId, err := cmd.Flags().GetString("id")
-	if err != nil {
-		return cmdError.NewCmdError(1, err.Error())
-	} else if volumeId == "" {
-		volumeList, err := exec.GetVolumeList(workspaceId)
-		if err != nil {
-			return cmdError.NewCmdError(1, err.Error())
-		}
-		printVolumeList(*volumeList)
-	} else {
-		volumeDetail, err := exec.GetVolume(workspaceId, volumeId)
-		if err != nil {
-			return cmdError.NewCmdError(1, err.Error())
-		}
-		printVolume(*volumeDetail)
-	}
-
-	return nil
 }
 
 func init() {
